feat(watchlist): add WatchlistSourceResultsCard and WatchlistSourcesCard components
- Implemented WatchlistSourceResultsCard to display job results from watchlist sources, including loading and error states.
- Created WatchlistSourcesCard for managing watchlist source drafts, allowing users to select catalog sources or add custom Workday URLs.
- Introduced types for watchlist functionality, including WatchlistFetchState, JobDetailsState, and RankedWorkdayJob.
- Added utility functions for handling watchlist sources and job ranking logic.
