+++ orchestrator/src/client/pages/watchlist/types.ts
@@ -1,0 +1,86 @@
+import type {
+  NormalizedWorkdayJob,
+  NormalizedWorkdayJobDetails,
+  WorkdayCxsJobsResult,
+} from "@client/api/workday";
+import type {
+  JobListItem,
+  ManualJobDraft,
+  WatchlistSelectedSource,
+  WatchlistSource,
+} from "@shared/types.js";
+
+export type WatchlistFetchState =
+  | {
+      status: "loading";
+      source: WatchlistSelectedSource;
+    }
+  | {
+      status: "success";
+      source: WatchlistSelectedSource;
+      response: WorkdayCxsJobsResult;
+    }
+  | {
+      status: "error";
+      source: WatchlistSelectedSource;
+      error: string;
+    };
+
+export interface SourceSelectionDraft {
+  id: string;
+  isCustom: boolean;
+  catalogSourceId: string | null;
+  customUrl: string;
+}
+
+export type JobDetailsState =
+  | {
+      status: "loading";
+    }
+  | {
+      status: "success";
+      details: NormalizedWorkdayJobDetails;
+    }
+  | {
+      status: "error";
+      error: string;
+    };
+
+export interface RankedWorkdayJob {
+  workdayJob: NormalizedWorkdayJob;
+  job: JobListItem;
+  matchScore: number;
+  matchedSearchTerm: string | null;
+  locationPriority: 0 | 1;
+  locationMatched: boolean;
+}
+
+export interface WorkdayImportState {
+  open: boolean;
+  draft: ManualJobDraft | null;
+  source: string | null;
+  sourceHost: string | null;
+}
+
+export type WatchlistRowState = "new" | "ignored" | "moved_to_workspace";
+
+export interface WatchlistCheckState {
+  checkedAt: string | null;
+  previousLastCheckedAt: string | null;
+  newJobKeys: Set<string>;
+}
+
+export interface WatchlistSourceDraftCardProps {
+  sourceDrafts: SourceSelectionDraft[];
+  catalogSources: WatchlistSource[];
+  formattedLastCheckedAt: string | null;
+  formattedPreviousLastCheckedAt: string | null;
+  newJobsCount: number;
+  isSaving: boolean;
+  onSourceCountChange: (nextCount: number) => void;
+  onUpdateDraft: (
+    index: number,
+    updater: (draft: SourceSelectionDraft) => SourceSelectionDraft,
+  ) => void;
+  onSave: () => void;
+}
